Re-name download package to client, concentrate HTTP calls there
