+++ pkg/alias/alias.go
@@ -2,6 +2,7 @@ package alias
 
 import (
 	"fmt"
+	"path/filepath"
 
 	"github.com/tfversion/tfversion/pkg/helpers"
 	"github.com/tfversion/tfversion/pkg/paths"
@@ -14,7 +15,7 @@ func AliasVersion(alias string, version string) {
 		helpers.ExitWithError("aliasing", err)
 	}
 
-	aliasPath := paths.GetAliasLocation(alias)
+	aliasPath := filepath.Join(paths.GetAliasLocation(), alias)
 	err := paths.RemoveSymlink(aliasPath)
 	if err != nil {
 		helpers.ExitWithError("removing symlink", err)
