+++ pkg/client/download.go
@@ -1,4 +1,4 @@
-package download
+package client
 
 import (
 	"fmt"
