--- pkg/list/const.go
@@ -1,6 +1,0 @@
-package list
-
-const (
-	// TerraformReleasesUrl is the URL used to list Terraform releases from.
-	TerraformReleasesUrl = "https://releases.hashicorp.com/terraform"
-)
